Adjust header files
