+++ librz/bin/pdb/omap.h
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2021 Basstorm <basstorm@nyist.edu.cn>
+// SPDX-License-Identifier: LGPL-3.0-only
+
+#ifndef PDB_OMAP_H
+#define PDB_OMAP_H
+
+#include <rz_util.h>
+
+typedef struct {
+	ut32 from;
+	ut32 to;
+} OmapEntry;
+
+#endif
